Adds import class and command
